fix: askforcreatorlabel migration error - 2

+++ auth_shop/migrations/0005_alter_askforcreatorlabel_options_and_more.py
@@ -17,6 +17,6 @@ class Migration(migrations.Migration):
         migrations.AlterField(
             model_name='askforcreatorlabel',
             name='asked_counter',
-            field=models.PositiveSmallIntegerField(default=1, verbose_name='Fois demandé'),
+            field=models.PositiveIntegerField(default=1, verbose_name='Fois demandé'),
         ),
     ]
